// src/app/interfaces/SignalElement.ts
export interface SignalElement {
  company: string;
  board_power: string;
  board_per_purchase_cost: string;
  board_number_purchases: string;
  recent_power: string;
  recent_per_purchase_cost: string;
  recent_number_purchases: string;
  now_power: string;
  now_per_purchase_cost: string;
  now_number_purchases: string;
  trade_price: string;
  final_price: string;
  opening_price: string;
  low_price: string;
  high_price: string;
  time: string;
  signal_type: string;
  url: string;
  score: string;
}